Fix: move pdb name check from Par to AdaptLoci
